app: use csv.DictWriter instead of csv.writer
csv.DictWriter allows specification of field names, and can
substitute empty values where the field value is missing. This
is useful in merging the endemic and non-endemic lists as the
endemic list has many missing fields and the order of the fields
does not necessarily match that of the non-endemic list.